refactor(styled-react): update structure to reduce merge conflicts
