+++ packages/styled-react/src/components/Box.tsx
@@ -1,0 +1,1 @@
+export {Box, type BoxProps} from '@primer/react'
