+++ packages/styled-react/src/components/StateLabel.tsx
@@ -1,0 +1,12 @@
+import {StateLabel as PrimerStateLabel, type StateLabelProps as PrimerStateLabelProps} from '@primer/react'
+import {forwardRef} from 'react'
+import {Box} from './Box'
+import type {SxProp} from '../sx'
+
+type StateLabelProps = PrimerStateLabelProps & SxProp
+
+const StateLabel = forwardRef<HTMLSpanElement, StateLabelProps>(function StateLabel(props, ref) {
+  return <Box as={PrimerStateLabel} ref={ref} {...props} />
+})
+
+export {StateLabel, type StateLabelProps}
