+++ packages/styled-react/src/components/SubNav.tsx
@@ -1,0 +1,26 @@
+import {
+  SubNav as PrimerSubNav,
+  type SubNavProps as PrimerSubNavProps,
+  type SubNavLinkProps as PrimerSubNavLinkProps,
+} from '@primer/react'
+import {forwardRef} from 'react'
+import {Box} from './Box'
+import type {SxProp} from '../sx'
+
+type SubNavProps = PrimerSubNavProps & SxProp
+
+const SubNavImpl = forwardRef<HTMLElement, SubNavProps>(function SubNav(props, ref) {
+  return <Box as={PrimerSubNav} ref={ref} {...props} />
+})
+
+type SubNavLinkProps = PrimerSubNavLinkProps & SxProp
+
+const SubNavLink = forwardRef<HTMLAnchorElement, SubNavLinkProps>(function SubNavLink(props, ref) {
+  return <Box as={PrimerSubNav.Link} ref={ref} {...props} />
+})
+
+const SubNav = Object.assign(SubNavImpl, {
+  Link: SubNavLink,
+})
+
+export {SubNav, type SubNavProps, type SubNavLinkProps}
