+++ packages/styled-react/src/index.tsx
@@ -1,96 +1,15 @@
-import {
-  type BetterSystemStyleObject,
-  Box,
-  type BoxProps,
-  type SxProp,
-  StateLabel as PrimerStateLabel,
-  type StateLabelProps as PrimerStateLabelProps,
-  SubNav as PrimerSubNav,
-  type SubNavProps as PrimerSubNavProps,
-  type SubNavLinkProps as PrimerSubNavLinkProps,
-  ToggleSwitch as PrimerToggleSwitch,
-  type ToggleSwitchProps as PrimerToggleSwitchProps,
-  type SegmentedControlProps as PrimerSegmentedControlProps,
-  SegmentedControl as PrimerSegmentedControl,
-  type SegmentedControlButtonProps as PrimerSegmentedControlButtonProps,
-  type SegmentedControlIconButtonProps as PrimerSegmentedControlIconButtonProps,
-} from '@primer/react'
-import React, {forwardRef, type PropsWithChildren} from 'react'
-import type {
-  BackgroundProps,
-  BorderProps,
-  ColorProps,
-  FlexboxProps,
-  GridProps,
-  LayoutProps,
-  PositionProps,
-  ShadowProps,
-  SpaceProps,
-  TypographyProps,
-} from 'styled-system'
-
-type StyledProps = SxProp &
-  SpaceProps &
-  ColorProps &
-  TypographyProps &
-  LayoutProps &
-  FlexboxProps &
-  GridProps &
-  BackgroundProps &
-  BorderProps &
-  PositionProps &
-  ShadowProps
-
-type SegmentedControlProps = PropsWithChildren<PrimerSegmentedControlProps> & SxProp
-type SegmentedControlButtonProps = PropsWithChildren<PrimerSegmentedControlButtonProps> & SxProp
-type SegmentedControlIconButtonProps = PropsWithChildren<PrimerSegmentedControlIconButtonProps> & SxProp
-
-const SegmentedControlButton = (props: SegmentedControlButtonProps) => {
-  return <Box as={PrimerSegmentedControl.Button} {...props} />
-}
-
-const SegmentedControlIconButton = (props: SegmentedControlIconButtonProps) => {
-  return <Box as={PrimerSegmentedControl.IconButton} {...props} />
-}
-
-const SegmentedControlImpl = (props: SegmentedControlProps) => {
-  return <Box as={PrimerSegmentedControl} {...props} />
-}
-
-const SegmentedControl = Object.assign(SegmentedControlImpl, {
-  Button: SegmentedControlButton,
-  IconButton: SegmentedControlIconButton,
-})
-
-type StateLabelProps = PrimerStateLabelProps & SxProp
-
-const StateLabel = forwardRef<HTMLSpanElement, StateLabelProps>(function StateLabel(props, ref) {
-  return <Box as={PrimerStateLabel} ref={ref} {...props} />
-})
-
-type SubNavProps = PrimerSubNavProps & SxProp
-
-const SubNavImpl = forwardRef<HTMLElement, SubNavProps>(function SubNav(props, ref) {
-  return <Box as={PrimerSubNav} ref={ref} {...props} />
-})
-
-type SubNavLinkProps = PrimerSubNavLinkProps & SxProp
-
-const SubNavLink = forwardRef<HTMLAnchorElement, SubNavLinkProps>(function SubNavLink(props, ref) {
-  return <Box as={PrimerSubNav.Link} ref={ref} {...props} />
-})
-
-const SubNav = Object.assign(SubNavImpl, {
-  Link: SubNavLink,
-})
-
-type ToggleSwitchProps = PrimerToggleSwitchProps & Omit<StyledProps, keyof PrimerToggleSwitchProps>
-
-const ToggleSwitch = forwardRef<HTMLButtonElement, ToggleSwitchProps>(function ToggleSwitch(props, ref) {
-  return <Box as={PrimerToggleSwitch} ref={ref} {...props} />
-})
-
-export {SegmentedControl, StateLabel, SubNav, ToggleSwitch}
+export {Box, type BoxProps} from './components/Box'
+export {
+  SegmentedControl,
+  type SegmentedControlProps,
+  type SegmentedControlButtonProps,
+  type SegmentedControlIconButtonProps,
+} from './components/SegmentedControl'
+export {SubNav, type SubNavProps, type SubNavLinkProps} from './components/SubNav'
+export {StateLabel, type StateLabelProps} from './components/StateLabel'
+export {ToggleSwitch, type ToggleSwitchProps} from './components/ToggleSwitch'
+export {sx, type SxProp} from './sx'
+export type {BetterSystemStyleObject} from './styled-props'
 
 export {
   ActionList,
@@ -131,10 +50,6 @@ export {
   Truncate,
   UnderlineNav,
 
-  // styled-components components or types
-  Box,
-  sx,
-
   // theming depends on styled-components
   ThemeProvider,
   merge,
@@ -143,4 +58,3 @@ export {
   useColorSchemeVar,
   useTheme,
 } from '@primer/react'
-export type {BoxProps, SxProp, BetterSystemStyleObject}
