+++ packages/styled-react/src/sx.ts
@@ -1,0 +1,1 @@
+export {sx, type SxProp} from '@primer/react'
